requirements.txt
fastapi
uvicorn[standard]
transformers
torch
sentencepiece
streamlit
requests
boto3
pymupdf
python-dotenv
python-multipart